Initial version of moving Sow with Size button to it's own section
